--- plugin/pulpcore/plugin/stages.py
@@ -1,783 +1,0 @@
-import asyncio
-from collections import defaultdict
-from gettext import gettext as _
-import os
-import shutil
-
-from django.db import transaction
-from django.db.models import Q
-
-from pulpcore.plugin.tasking import WorkingDirectory
-from pulpcore.plugin.models import (
-    Artifact, ContentArtifact, ProgressBar, RemoteArtifact, RepositoryVersion
-)
-
-
-class DeclarativeArtifact:
-    """
-    Relates an Artifact, how to download it, and its relative_path for publishing.
-
-    This is used by the Stages API stages to determine if an Artifact is already present and ensure
-    Pulp can download it in the future. The `artifact` can be either saved or unsaved. If unsaved,
-    the `artifact` attributes may be incomplete because not all digest information can be computed
-    until the Artifact is downloaded.
-
-    Attributes:
-        artifact - An Artifact either saved or unsaved. If unsaved, it may have partial digest
-            information attached to it.
-        url - the url to fetch the Artifact from.
-        relative_path - the relative_path this Artifact should be published at for any Publication.
-        remote - The remote used to fetch this Artifact.
-
-    Raises:
-        ValueError: If `artifact`, `url`, `relative_path`, or `remote` are not specified.
-    """
-
-    __slots__ = ('artifact', 'url', 'relative_path', 'remote')
-
-    def __init__(self, artifact=None, url=None, relative_path=None, remote=None):
-        if not url:
-            raise ValueError(_("DeclarativeArtifact must have a 'url'"))
-        if not relative_path:
-            raise ValueError(_("DeclarativeArtifact must have a 'relative_path'"))
-        if not remote:
-            raise ValueError(_("DeclarativeArtifact must have a 'remote'"))
-        if not artifact:
-            raise ValueError(_("DeclarativeArtifact must have a 'artifact'"))
-        self.artifact = artifact
-        self.url = url
-        self.relative_path = relative_path
-        self.remote = remote
-
-
-class DeclarativeContent:
-    """
-    Relates a Content unit and zero or more DeclarativeArtifact objects.
-
-    This is used by the Stages API stages to determine if a Content unit is already present and
-    ensure all of its associated DeclarativeArtifact objects are related correctly. The `content`
-    can be either saved or unsaved depending on where in the Stages API pipeline this is used.
-
-    Attributes:
-        content - The in-memory, partial Artifact with any known digest information attached to it
-        d_artifacts - A list of zero or more DeclarativeArtifacts associated with `content`.
-
-    Raises:
-        ValueError: If `content` is not specified.
-    """
-
-    __slots__ = ('content', 'd_artifacts')
-
-    def __init__(self, content=None, d_artifacts=None):
-        if not content:
-            raise ValueError(_("DeclarativeContent must have a 'content'"))
-        if d_artifacts:
-            self.d_artifacts = d_artifacts
-        else:
-            self.d_artifacts = []
-        self.content = content
-
-
-async def create_pipeline(stages, in_q=None, maxsize=100):
-    """
-    Creates a Stages API linear pipeline from the list `stages` and return as a single coroutine.
-
-    Each stage is a coroutine and reads from an input `asyncio.Queue` and writes to an output
-    `asyncio.Queue`. When the stage is ready to shutdown it writes a `None` to the output Queue.
-    Here is an example of the simplest stage that only passes data.
-
-    >>> async def my_stage(in_q, out_q):
-    >>>     while True:
-    >>>         item = await in_q.get()
-    >>>         if item is None:  # Check if the previous stage is shutdown
-    >>>             break
-    >>>         await out_q.put(item)
-    >>>     await out_q.put(None)  # this stage is shutdown so send 'None'
-
-    Args:
-        stages (list of coroutines): A list of Stages API compatible coroutines.
-        in_q (asyncio.Queue): The queue the first stage should read from. This is how to put work
-            into the pipeline. Optional especially for cases where the first stage generates items
-            for `out_q` without needing inputs from `in_q`.
-        maxsize (int): The maximum amount of items a queue between two stages should hold. Optional
-            and defaults to 100.
-
-    Returns:
-        A single coroutine that can be used to run, wait, or cancel the entire pipeline with.
-    """
-    futures = []
-    for stage in stages:
-        out_q = asyncio.Queue(maxsize=maxsize)
-        futures.append(stage(in_q, out_q))
-        in_q = out_q
-    await asyncio.gather(*futures)
-
-
-class FirstStage:
-    """
-    A class plugin writers can subclass and use as the first stage for a DeclarativeVersion pipeline
-
-    To use this class, the plugin writer needs to:
-
-    1. Subclass it and implement the `gen_declarative_content()` method.
-    2. Pass the instantiated subclass to `DeclarativeVersion`.
-    """
-
-    async def gen_declarative_content(self, in_q, out_q):
-        """
-        A Stages API compatible coroutine for `DeclarativeVersion` to use as the first stage.
-
-        This must be implemented on the subclass.
-
-        Args:
-            in_q (asyncio.Queue): Unused because the first stage doesn't read from an input queue.
-            out_q (asyncio.Queue): The queue to put `DeclarativeContent` into.
-
-        Returns:
-            A Stages API compatible coroutine for `DeclarativeVersion` to use as the first stage.
-        """
-        raise NotImplementedError('A plugin writer needs to implement this')
-
-
-async def query_existing_artifacts(in_q, out_q):
-    """
-    Stages API stage replacing DeclarativeArtifact.artifact with already-saved Artifacts.
-
-    This stage expects `~pulpcore.plugin.stages.DeclarativeContent` units from `in_q` and inspects
-    their associated `~pulpcore.plugin.stages.DeclarativeArtifact` objects. Each
-    `DeclarativeArtifact` object stores one Artifact.
-
-    This stage inspects any "unsaved" Artifact objects metadata and searches for existing saved
-    Artifacts inside Pulp with the same digest value(s). Any existing Artifact objects found replace
-    their "unsaved" counterpart in the `~pulpcore.plugin.stages.DeclarativeArtifact` object.
-
-    Each `~pulpcore.plugin.stages.DeclarativeContent` is sent to `out_q` after all of its
-    `~pulpcore.plugin.stages.DeclarativeArtifact` objects have been handled.
-
-    This stage drains all available items from `in_q` and batches everything into one large call to
-    the db for efficiency.
-
-    Args:
-        in_q: `~pulpcore.plugin.stages.DeclarativeContent`
-        out_q: `~pulpcore.plugin.stages.DeclarativeContent`
-
-    Returns:
-        The query_existing_artifacts stage as a coroutine to be included in a pipeline.
-    """
-    declarative_content = []
-    shutdown = False
-    while True:
-        try:
-            content = in_q.get_nowait()
-        except asyncio.QueueEmpty:
-            if not declarative_content:
-                content = await in_q.get()
-                declarative_content.append(content)
-                continue
-        else:
-            declarative_content.append(content)
-            continue
-
-        all_artifacts_q = Q(pk=None)
-        for content in declarative_content:
-            if content is None:
-                shutdown = True
-                continue
-
-            for declarative_artifact in content.d_artifacts:
-                one_artifact_q = Q()
-                for digest_name in declarative_artifact.artifact.DIGEST_FIELDS:
-                    digest_value = getattr(declarative_artifact.artifact, digest_name)
-                    if digest_value:
-                        key = {digest_name: digest_value}
-                        one_artifact_q &= Q(**key)
-                all_artifacts_q |= one_artifact_q
-
-        for artifact in Artifact.objects.filter(all_artifacts_q):
-            for content in declarative_content:
-                if content is None:
-                    continue
-                for declarative_artifact in content.d_artifacts:
-                    for digest_name in artifact.DIGEST_FIELDS:
-                        digest_value = getattr(declarative_artifact.artifact, digest_name)
-                        if digest_value and digest_value == getattr(artifact, digest_name):
-                            declarative_artifact.artifact = artifact
-        for content in declarative_content:
-            if content is None:
-                continue
-            await out_q.put(content)
-        declarative_content = []
-        if shutdown:
-            break
-    await out_q.put(None)
-
-
-class artifact_downloader:
-    """
-    An object containing a Stages API stage to download missing Artifacts, but don't call save()
-
-    The actual stage is the `stage` attribute which can be used as follows:
-
-    >>> artifact_downloader(max_concurrent_downloads=42).stage  # This is the real stage
-
-    in_q data type: A `~pulpcore.plugin.stages.DeclarativeContent` with potentially files missing
-        `~pulpcore.plugin.stages.DeclarativeArtifact.artifact` objects.
-    out_q data type: A `~pulpcore.plugin.stages.DeclarativeContent` with all files downloaded for
-        all `~pulpcore.plugin.stages.DeclarativeArtifact.artifact` objects.
-
-    This stage downloads the file for any Artifact objects missing files and creates a new Artifact
-    from the downloaded file and its digest data. The new Artifact is *not* saved but added to the
-    `~pulpcore.plugin.stages.DeclarativeArtifact` object, replacing the likely incomplete Artifact.
-
-    Each `~pulpcore.plugin.stages.DeclarativeContent` is sent to `out_q` after all of its
-    `~pulpcore.plugin.stages.DeclarativeArtifact` objects have been handled.
-
-    This stage creates a ProgressBar named 'Downloading Artifacts' that counts the number of
-    downloads completed. Since it's a stream the total count isn't known until it's finished.
-
-    This stage drains all available items from `in_q` and starts as many downloaders as possible.
-
-    Args:
-        max_concurrent_downloads (int): The maximum number of concurrent downloads this stage will
-            run. Default is 100.
-
-    Returns:
-        An object containing the artifact_downloader stage to be included in a pipeline.
-    """
-
-    def __init__(self, max_concurrent_downloads=100):
-        self.max_concurrent_downloads = max_concurrent_downloads
-
-    async def stage(self, in_q, out_q):
-        """Download undownloaded Artifacts, but don't save them"""
-        pending = set()
-        incoming_content = []
-        outstanding_downloads = 0
-        shutdown = False
-        saturated = False
-        with ProgressBar(message='Downloading Artifacts') as pb:
-            while True:
-                if not saturated:
-                    try:
-                        content = in_q.get_nowait()
-                    except asyncio.QueueEmpty:
-                        if not incoming_content and not shutdown and not pending:
-                            content = await in_q.get()
-                            incoming_content.append(content)
-                            continue
-                    else:
-                        incoming_content.append(content)
-                        continue
-
-                for i, content in enumerate(incoming_content):
-                    if content is None:
-                        shutdown = True
-                        continue
-                    downloaders_for_content = []
-                    for declarative_artifact in content.d_artifacts:
-                        if declarative_artifact.artifact._state.adding:
-                            # this needs to be downloaded
-                            downloader = declarative_artifact.remote.get_downloader(
-                                declarative_artifact.url
-                            )
-                            next_future = asyncio.ensure_future(downloader.run())
-                            downloaders_for_content.append(next_future)
-                    if not downloaders_for_content:
-                        await out_q.put(content)
-                        continue
-
-                    async def return_content_for_downloader(c):
-                        return c
-                    outstanding_downloads = outstanding_downloads + len(downloaders_for_content)
-                    downloaders_for_content.append(return_content_for_downloader(content))
-                    pending.add(asyncio.gather(*downloaders_for_content))
-                    if outstanding_downloads > self.max_concurrent_downloads:
-                        saturated = True
-                        incoming_content = incoming_content[i + 1:]  # remove handled content
-                        break
-                else:
-                    incoming_content = []
-
-                if pending:
-                    done, pending = await asyncio.wait(pending, return_when=asyncio.FIRST_COMPLETED)
-                    for gathered_downloaders in done:
-                        results = gathered_downloaders.result()
-                        for download_result in results[:-1]:
-                            content = results[-1]
-                            for declarative_artifact in content.d_artifacts:
-                                if declarative_artifact.artifact._state.adding:
-                                    new_artifact = Artifact(
-                                        **download_result.artifact_attributes,
-                                        file=download_result.path
-                                    )
-                                    declarative_artifact.artifact = new_artifact
-                            pb.done = pb.done + len(content.d_artifacts)
-                            outstanding_downloads = outstanding_downloads - len(content.d_artifacts)
-                            await out_q.put(content)
-                else:
-                    if shutdown:
-                        break
-
-                if outstanding_downloads < self.max_concurrent_downloads:
-                    saturated = False
-        await out_q.put(None)
-
-
-async def artifact_saver(in_q, out_q):
-    """
-    Stages API stage that saves an Artifact for any "unsaved" DeclarativeArtifact.artifact objects.
-
-    This stage expects `~pulpcore.plugin.stages.DeclarativeContent` units from `in_q` and inspects
-    their associated `~pulpcore.plugin.stages.DeclarativeArtifact` objects. Each
-    `DeclarativeArtifact` object stores one Artifact.
-
-    Any "unsaved" Artifact objects are saved. Each `~pulpcore.plugin.stages.DeclarativeContent` is
-    sent to `out_q` after all of its `~pulpcore.plugin.stages.DeclarativeArtifact` objects have been
-    handled.
-
-    Args:
-        in_q: `~pulpcore.plugin.stages.DeclarativeContent`
-        out_q: `~pulpcore.plugin.stages.DeclarativeContent`
-
-    Returns:
-        The artifact_saver stage as a coroutine to be included in a pipeline.
-    """
-    shutdown = False
-    declarative_content_list = []
-    while not shutdown:
-        try:
-            content = in_q.get_nowait()
-        except asyncio.QueueEmpty:
-            if not declarative_content_list:
-                content = await in_q.get()
-                declarative_content_list.append(content)
-                continue
-        else:
-            declarative_content_list.append(content)
-            continue
-
-        artifacts_to_save = []
-        for declarative_content in declarative_content_list:
-            if declarative_content is None:
-                shutdown = True
-                break
-            for declarative_artifact in declarative_content.d_artifacts:
-                if declarative_artifact.artifact._state.adding:
-                    src_path = str(declarative_artifact.artifact.file)
-                    dst_path = declarative_artifact.artifact.storage_path(None)
-                    try:
-                        shutil.move(src_path, dst_path)
-                    except FileNotFoundError:
-                        os.makedirs(os.path.dirname(dst_path), exist_ok=True)
-                        shutil.move(src_path, dst_path)
-                    declarative_artifact.artifact.file = dst_path
-                    artifacts_to_save.append(declarative_artifact.artifact)
-
-        if artifacts_to_save:
-            Artifact.objects.bulk_create(artifacts_to_save)
-
-        for declarative_content in declarative_content_list:
-            if declarative_content is None:
-                continue
-            await out_q.put(declarative_content)
-
-        declarative_content_list = []
-
-    await out_q.put(None)
-
-
-async def query_existing_content_units(in_q, out_q):
-    """
-    Stages API stage replacing DeclarativeContent.content with already-saved Content objects.
-
-    This stage expects `~pulpcore.plugin.stages.DeclarativeContent` units from `in_q` and inspects
-    their associated `~pulpcore.plugin.stages.DeclarativeArtifact` objects. Each
-    `DeclarativeArtifact` object stores one Artifact.
-
-    This stage inspects any "unsaved" Content unit objects and searches for existing saved Content
-    units inside Pulp with the same unit key. Any existing Content objects found replace their
-    "unsaved" counterpart in the `~pulpcore.plugin.stages.DeclarativeContent` object.
-
-    Each `~pulpcore.plugin.stages.DeclarativeContent` is sent to `out_q` after it has been handled.
-
-    This stage drains all available items from `in_q` and batches everything into one large call to
-    the db for efficiency.
-
-    Args:
-        in_q: `~pulpcore.plugin.stages.DeclarativeContent`
-        out_q: `~pulpcore.plugin.stages.DeclarativeContent`
-
-    Returns:
-        The query_existing_content_units stage as a coroutine to be included in a pipeline.
-    """
-    declarative_content_list = []
-    shutdown = False
-    while True:
-        try:
-            content = in_q.get_nowait()
-        except asyncio.QueueEmpty:
-            if not declarative_content_list:
-                content = await in_q.get()
-                declarative_content_list.append(content)
-                continue
-        else:
-            declarative_content_list.append(content)
-            continue
-
-        content_q_by_type = defaultdict(lambda: Q(pk=None))
-        for declarative_content in declarative_content_list:
-            if declarative_content is None:
-                shutdown = True
-                continue
-
-            model_type = type(declarative_content.content)
-            unit_key = declarative_content.content.natural_key_dict()
-            content_q_by_type[model_type] = content_q_by_type[model_type] | Q(**unit_key)
-
-        for model_type in content_q_by_type.keys():
-            for result in model_type.objects.filter(content_q_by_type[model_type]):
-                for declarative_content in declarative_content_list:
-                    if declarative_content is None:
-                        continue
-                    not_same_unit = False
-                    for field in result.natural_key_fields():
-                        if getattr(declarative_content.content, field) != getattr(result, field):
-                            not_same_unit = True
-                            break
-                    if not_same_unit:
-                        continue
-                    declarative_content.content = result
-        for declarative_content in declarative_content_list:
-            await out_q.put(declarative_content)
-        declarative_content_list = []
-        if shutdown:
-            break
-    await out_q.put(None)
-
-
-async def content_unit_saver(in_q, out_q):
-    """
-    Stages API stage that saves DeclarativeContent.content objects and saves helper objects too.
-
-    This stage expects `~pulpcore.plugin.stages.DeclarativeContent` units from `in_q` and inspects
-    their associated `~pulpcore.plugin.stages.DeclarativeArtifact` objects. Each
-    `DeclarativeArtifact` object stores one Artifact.
-
-    Each "unsaved" Content objects is saved and a RemoteArtifact and ContentArtifact are created for
-    it and saved also. This allows Pulp to refetch the Artifact in the future if the local copy is
-    removed.
-
-    Each `~pulpcore.plugin.stages.DeclarativeContent` is sent to after it has been handled.
-
-    This stage drains all available items from `in_q` and batches everything into one large call to
-    the db for efficiency.
-
-    Args:
-        in_q: `~pulpcore.plugin.stages.DeclarativeContent`
-        out_q: `~pulpcore.plugin.stages.DeclarativeContent`
-
-    Returns:
-        The content_unit_saver stage as a coroutine to be included in a pipeline.
-    """
-    declarative_content_list = []
-    shutdown = False
-    while True:
-        try:
-            declarative_content = in_q.get_nowait()
-        except asyncio.QueueEmpty:
-            if not declarative_content_list and not shutdown:
-                declarative_content = await in_q.get()
-                declarative_content_list.append(declarative_content)
-                continue
-        else:
-            declarative_content_list.append(declarative_content)
-            continue
-
-        content_artifact_bulk = []
-        remote_artifact_bulk = []
-        remote_artifact_map = {}
-
-        with transaction.atomic():
-            for declarative_content in declarative_content_list:
-                if declarative_content is None:
-                    shutdown = True
-                    continue
-                if declarative_content.content._state.adding:
-                        declarative_content.content.save()
-                        for declarative_artifact in declarative_content.d_artifacts:
-                            content_artifact = ContentArtifact(
-                                content=declarative_content.content,
-                                artifact=declarative_artifact.artifact,
-                                relative_path=declarative_artifact.relative_path
-                            )
-                            content_artifact_bulk.append(content_artifact)
-                            remote_artifact_data = {
-                                'url': declarative_artifact.url,
-                                'size': declarative_artifact.artifact.size,
-                                'md5': declarative_artifact.artifact.md5,
-                                'sha1': declarative_artifact.artifact.sha1,
-                                'sha224': declarative_artifact.artifact.sha224,
-                                'sha256': declarative_artifact.artifact.sha256,
-                                'sha384': declarative_artifact.artifact.sha384,
-                                'sha512': declarative_artifact.artifact.sha512,
-                                'remote': declarative_artifact.remote,
-                            }
-                            rel_path = content_artifact.relative_path
-                            remote_artifact_map[rel_path] = remote_artifact_data
-
-            for content_artifact in ContentArtifact.objects.bulk_create(content_artifact_bulk):
-                remote_artifact_data = remote_artifact_map.pop(content_artifact.relative_path)
-                new_remote_artifact = RemoteArtifact(
-                    content_artifact=content_artifact, **remote_artifact_data
-                )
-                remote_artifact_bulk.append(new_remote_artifact)
-
-            RemoteArtifact.objects.bulk_create(remote_artifact_bulk)
-
-        for declarative_content in declarative_content_list:
-            if declarative_content is None:
-                continue
-            await out_q.put(declarative_content)
-        if shutdown:
-            break
-        declarative_content_list = []
-    await out_q.put(None)
-
-
-class content_unit_association:
-    """
-    An object containing a Stages API stage that associates content units with `new_version`.
-
-    The actual stage is the `stage` attribute which can be used as follows:
-
-    >>> content_unit_association(my_new_version).stage  # This is the real stage
-
-    This stage stores all content unit types and unit keys in memory before running. This is done to
-    compute the units already associated but not received from `in_q`. These units are passed via
-    `out_q` to the next stage as a `~django.db.models.query.QuerySet`.
-
-    in_q data type: A `~pulpcore.plugin.stages.DeclarativeContent` with saved `content` that needs
-        to be associated.
-    out_q data type: A `django.db.models.query.QuerySet` of `~pulpcore.plugin.models.Content` or
-        subclass that are already associated but not included in the stream of items from `in_q`.
-        One `django.db.models.query.QuerySet` is put for each `~pulpcore.plugin.models.Content`
-        type.
-
-    This stage creates a ProgressBar named 'Associating Content' that counts the number of units
-    associated. Since it's a stream the total count isn't known until it's finished.
-
-    Args:
-        new_version (RepositoryVersion): The RespositoryVersion this stage associates content with.
-
-    Returns:
-        An object containing the content_unit_association stage to be included in a pipeline.
-    """
-
-    def __init__(self, new_version):
-        self.new_version = new_version
-        self.unit_keys_by_type = defaultdict(set)
-        for unit in self.new_version.content.all():
-            unit = unit.cast()
-            self.unit_keys_by_type[type(unit)].add(unit.natural_key())
-
-    async def stage(self, in_q, out_q):
-        """For each Content Unit associate it with the repository version"""
-        with ProgressBar(message='Associating Content') as pb:
-            declarative_content_list = []
-            shutdown = False
-            while True:
-                try:
-                    declarative_content = in_q.get_nowait()
-                except asyncio.QueueEmpty:
-                    if not declarative_content_list and not shutdown:
-                        declarative_content = await in_q.get()
-                        declarative_content_list.append(declarative_content)
-                        continue
-                else:
-                    declarative_content_list.append(declarative_content)
-                    continue
-
-                content_q_by_type = defaultdict(lambda: Q(pk=None))
-                for declarative_content in declarative_content_list:
-                    if declarative_content is None:
-                        shutdown = True
-                        continue
-                    try:
-                        unit_key = declarative_content.content.natural_key()
-                        self.unit_keys_by_type[type(declarative_content.content)].remove(unit_key)
-                    except KeyError:
-                        model_type = type(declarative_content.content)
-                        unit_key_dict = declarative_content.content.natural_key_dict()
-                        unit_q = Q(**unit_key_dict)
-                        content_q_by_type[model_type] = content_q_by_type[model_type] | unit_q
-
-                for model_type, q_object in content_q_by_type.items():
-                    queryset = model_type.objects.filter(q_object)
-                    self.new_version.add_content(queryset)
-                    pb.done = pb.done + queryset.count()
-                    pb.save()
-
-                if shutdown:
-                    break
-                declarative_content_list = []
-
-            for unit_type, ids in self.unit_keys_by_type.items():
-                if ids:
-                    units_to_unassociate = Q()
-                    for unit_key in self.unit_keys_by_type[unit_type]:
-                        query_dict = {}
-                        for i, key_name in enumerate(unit_type.natural_key_fields()):
-                            query_dict[key_name] = unit_key[i]
-                        units_to_unassociate |= Q(**query_dict)
-                    await out_q.put(unit_type.objects.filter(units_to_unassociate))
-            await out_q.put(None)
-
-
-class content_unit_unassociation:
-    """
-    An object containing a Stages API stage that unassociates content units from `new_version`.
-
-    The actual stage is the `stage` attribute which can be used as follows:
-
-    >>> content_unit_unassociation(my_new_version).stage  # This is the real stage
-
-    in_q data type: `django.db.models.query.QuerySet` of `~pulpcore.plugin.models.Content` or
-        subclass to be unassociated from `new_version`.
-    out_q data type: `django.db.models.query.QuerySet` of `~pulpcore.plugin.models.Content` or
-        subclass that were unassociated from `new_version`.
-
-    This stage creates a ProgressBar named 'Un-Associating Content' that counts the number of units
-    un-associated. Since it's a stream the total count isn't known until it's finished.
-
-    Args:
-        new_version (RepositoryVersion): The RespositoryVersion this stage unassociates content from
-
-    Returns:
-        An object containing the configured content_unit_unassociation stage to be included in a
-            pipeline.
-    """
-
-    def __init__(self, new_version):
-        self.new_version = new_version
-
-    async def stage(self, in_q, out_q):
-            """For each Content Unit from in_q, unassociate it with the repository version"""
-            with ProgressBar(message='Un-Associating Content') as pb:
-                while True:
-                    queryset_to_unassociate = await in_q.get()
-                    if queryset_to_unassociate is None:
-                        break
-
-                    self.new_version.remove_content(queryset_to_unassociate)
-                    pb.done = pb.done + queryset_to_unassociate.count()
-                    pb.save()
-
-                    await out_q.put(queryset_to_unassociate)
-                await out_q.put(None)
-
-
-async def end_stage(in_q, out_q):
-    """
-    A Stages API stage that drains `in_q` and do nothing with the items. This is expected at the end
-    of all pipelines.
-
-    Without this stage, the maxsize of the `out_q` from the last stage could fill up and block the
-    entire pipeline.
-    """
-    while True:
-        content = await in_q.get()
-        if content is None:
-            break
-
-
-class DeclarativeVersion:
-
-    def __init__(self, first_stage, repository, sync_mode='mirror'):
-        """
-        A pipeline that creates a new RepositoryVersion from a stream of DeclarativeContent objects.
-
-        The plugin writer needs to specify a first_stage that will create a
-        `~pulpcore.plugin.stages.DeclarativeContent` object for each Content unit that should exist
-        in the new RepositoryVersion.
-
-        The pipeline stages perform the following steps:
-
-        1. Create the new RespositoryVersion
-        2. Query existing artifacts to determine which are already local to Pulp
-        3. Download the undownloaded Artifacts
-        4. Save the newly downloaded Artifacts
-        5. Query for content units already present in Pulp
-        6. Save new content units not yet present in Pulp
-        7. Associate all content units with the new repository version.
-        8. Unassociate any content units not declared in the stream (only when sync_mode='mirror')
-
-        To do this, the plugin writer should subclass the FirstStage class and define its
-        `gen_declarative_content()` interface which return a coroutine. This coroutine should
-        download metadata, create the corresponding DeclarativeContent objects, and put them into
-        the `asyncio.Queue` to send them down the pipeline. For example:
-
-        >>> class MyFirstStage(FirstStage):
-        >>>
-        >>>     def __init__(remote):
-        >>>         self.remote = remote
-        >>>
-        >>>     async def gen_declarative_content(self, out_q):
-        >>>         downloader = remote.get_downloader(remote.url)
-        >>>         result = await downloader.run()
-        >>>         for entry in read_my_metadata_file_somehow(result.path)
-        >>>             unit = MyContent(entry)  # make the content unit in memory-only
-        >>>             artifact = Artifact(entry)  # make Artifact in memory-only
-        >>>             da = DeclarativeArtifact(artifact, url, entry.relative_path, self.remote)
-        >>>             dc = DeclarativeContent(content=unit, d_artifacts=[da])
-        >>>             await out_q.put(dc)
-        >>>         await out_q.put(None)
-
-        To use your first stage with the pipeline you have to instantiate the subclass and pass it
-        to `DeclarativeVersion`.
-
-        1. Create the instance of the FirstStage subclass
-        2. Create the `DeclarativeVersion` instance, passing the FirstStage subclass instance to it
-        3. Call the `create()` method on your `DeclarativeVersion` instance
-
-        Here is an example:
-
-        >>> first_stage = FileFirstStage(remote)
-        >>> DeclarativeVersion(first_stage, repository).create()
-
-        Args:
-             first_stage (FirstStage): The first stage to receive `DeclarativeContent` from
-             repository (Repository): The repository receiving the new version
-             sync_mode (str): 'mirror' removes content units from the RepositoryVersion that are not
-                 queued to DeclarativeVersion. 'additive' only adds content units queued to
-                 DeclarativeVersion, and does not remove any pre-existing units in the
-                 RepositoryVersion. 'mirror' is the default.
-
-        Raises:
-            ValueError: if 'sync_mode' is passed an invalid value.
-        """
-        if sync_mode is not 'mirror' and sync_mode is not 'additive':
-            msg = _("'sync_mode' must either be 'mirror' or 'additive' not '{sync_mode}'")
-            raise ValueError(msg.format(sync_mode=sync_mode))
-        self.first_stage = first_stage
-        self.repository = repository
-        self.sync_mode = sync_mode
-
-    def create(self):
-        """
-        Perform the work. This is the long-blocking call where all syncing occurs.
-        """
-        with WorkingDirectory():
-            with RepositoryVersion.create(self.repository) as new_version:
-                loop = asyncio.get_event_loop()
-                stages = [
-                    self.first_stage.gen_declarative_content,
-                    query_existing_artifacts, artifact_downloader().stage, artifact_saver,
-                    query_existing_content_units, content_unit_saver,
-                    content_unit_association(new_version).stage
-                ]
-                if self.sync_mode is 'additive':
-                    stages.append(end_stage)
-                elif self.sync_mode is 'mirror':
-                    stages.extend([content_unit_unassociation(new_version).stage, end_stage])
-                pipeline = create_pipeline(stages)
-                loop.run_until_complete(pipeline)
